Split test case agent helpers by module (#41)
