+++ docs/codebase/CONCERNS.md
@@ -7,7 +7,7 @@ git/history checks. It is not a full bug backlog.
 
 | Severity | Concern | Evidence | Impact | Suggested action |
 |----------|---------|----------|--------|------------------|
-| High | Large, high-churn orchestration files remain central to the app | `frontend/src/App.jsx`, `backend/app/agents/test_case_agent.py`, `backend/app/models.py`, recent git history | Higher regression and merge risk for UI workflow and generation behavior | Continue behavior-preserving extraction under issue-scoped frontend and agent refactor tasks |
+| High | Large, high-churn orchestration and contract files remain central to the app | `frontend/src/App.jsx`, `backend/app/agents/test_case_agent.py`, `backend/app/models.py`, recent git history | Higher regression and merge risk for UI workflow, prompt orchestration, and API contracts | Continue behavior-preserving extraction under issue-scoped frontend, agent, and contract refactor tasks |
 | High | Persistence target is not fully settled in docs versus code | `docs/firebase-auth-audit-architecture.md`, `backend/app/services/firebase_admin.py`, `backend/app/services/billing_repository.py` | Architecture decisions can drift between Firestore current state and Postgres target recommendations | Decide whether Firestore remains the near-term durable store or whether Postgres migration is active |
 | Medium | Dual auth support can blur production policy | `backend/app/auth/jwt_auth.py`, `backend/app/auth/firebase_auth.py`, `backend/app/routers/auth.py`, `frontend/src/App.jsx` | Local/E2E JWT compatibility is useful, but production paths need clear accepted-token policy | Document deployment auth mode and eventually remove or isolate legacy JWT if no longer needed |
 | Medium | `.env.example` duplicates execution settings | `.env.example` lines 34-50 | Increases setup confusion and risk of stale values | Deduplicate the repeated execution block in a focused docs/config cleanup issue |
@@ -21,7 +21,7 @@ git/history checks. It is not a full bug backlog.
 |-----------|---------------|-------|-----------------|---------------|
 | Monolithic frontend orchestration | `App.jsx` still owns most workflow state and API actions after component extraction | `frontend/src/App.jsx` | Harder UI changes and fragile E2E updates | Extract workflow state into hooks/reducers in small issue-scoped slices |
 | Shared frontend style cascade | Feature styles are split into owned modules but still share one imported cascade | `frontend/src/styles/index.css`, `frontend/src/styles/*.css` | Cross-feature selectors and import-order changes can still cause visual regressions | Keep new selectors in the closest feature style file and cover visual workflow changes with browser checks |
-| Large generation agent | Test-case generation, review, metrics, fallback, normalization, and prompting still share one module | `backend/app/agents/test_case_agent.py` | Agent behavior changes are difficult to isolate | Split orchestration, normalization, coverage metrics, fallback, and prompt helpers gradually |
+| Test-case agent orchestration | Coverage metrics, review scoring, fallback generation, and hydration are split, but prompt builders and workflow orchestration still share one module | `backend/app/agents/test_case_agent.py`, `backend/app/agents/test_case_coverage.py`, `backend/app/agents/test_case_review.py`, `backend/app/agents/test_case_fallback.py`, `backend/app/agents/test_case_hydration.py` | Prompt or workflow changes still require careful regression checks across generation behavior | Keep future prompt/orchestration changes focused and rely on helper-level tests plus offline benchmarks |
 | Broad model module | Many product domains share one Pydantic file | `backend/app/models.py` | Contract merge conflicts and long review cycles | Split model modules only after router/service ownership boundaries are stable |
 | Older roadmap/doc drift | Some docs describe work as future that is partly implemented now | `docs/implementation-plan.md`, `docs/frontend-refactor-github-issues.md` | Contributors may follow stale sequencing | Add "current state" notes or close superseded plan sections during planning updates |
 | No generated frontend API types | Frontend manually consumes backend response shapes | `frontend/src/App.jsx`, `scripts/export_openapi.py` | Response-shape drift can escape until runtime/E2E | Generate frontend types from OpenAPI in a dedicated contract issue |
@@ -52,7 +52,7 @@ git/history checks. It is not a full bug backlog.
 |------|-------------|--------------|----------------------|
 | `frontend/src/App.jsx` | Owns auth, workflow state, API actions, derived metrics, and layout orchestration | Highest recent git churn among tracked source hotspots; 3013 lines | Keep changes focused and cover with frontend build/E2E |
 | `frontend/src/styles/` | Feature-owned stylesheet modules imported through a shared cascade | Former `App.css` styles now span feature files and still depend on import order | Keep selector moves scoped and verify primary workflow screens visually |
-| `backend/app/agents/test_case_agent.py` | Central generation/review/coverage/fallback pipeline | High churn and 3245 lines | Add focused tests for every behavior slice before refactor |
+| `backend/app/agents/test_case_agent.py` and focused test-case helper modules | Central generation orchestration plus split coverage, review, fallback, and hydration helpers | High churn; orchestrator reduced to about 1400 lines with helpers in `test_case_*.py` modules | Add focused tests for every behavior slice before changing prompt or workflow logic |
 | `backend/app/models.py` | Shared API contracts for many domains | High churn and 1047 lines | Treat changes as contract changes; run OpenAPI export |
 | Integration sync routers/services | JIRA and Azure DevOps have parallel flows with provider-specific edge cases | Recent traceability and test isolation work | Change one provider at a time or keep explicit parity tests |
 | Execution runtime | Converts natural language cases into runnable browser specs | New next-version E2E capability | Run execution preview and runtime checks for every change |
@@ -80,4 +80,8 @@ git/history checks. It is not a full bug backlog.
 - `frontend/src/App.jsx`
 - `frontend/src/styles/`
 - `backend/app/agents/test_case_agent.py`
+- `backend/app/agents/test_case_coverage.py`
+- `backend/app/agents/test_case_review.py`
+- `backend/app/agents/test_case_fallback.py`
+- `backend/app/agents/test_case_hydration.py`
 - `backend/app/models.py`
