+++ docs/requirements_traceability.md
@@ -14,3 +14,4 @@ This matrix links requirement issues to implementation, tests, and validation ev
 | [#22](https://github.com/lumensparkxy/agentic_test_case_generator/issues/22) | Isolate backend sync and audit endpoint tests from Firestore and billing side effects that are outside the focused behavior under test. | `backend/app/services/jira_sync_service.py`, `backend/app/services/azure_devops_sync_service.py`, `backend/tests/test_jira_sync_service.py`, `backend/tests/test_azure_devops_sync_service.py`, `backend/tests/test_main_audit_logging.py` | JIRA and Azure DevOps sync service tests assert direct source metadata avoids Firestore mapping reads; audit logging endpoint tests patch and assert billing access and consumption boundaries. | Focused #22 tests passed 7 tests; backend suite passed 191 tests; strict requirement evaluation passed at 100/85; strict generation evaluation passed at 100/90; `scripts/export_openapi.py` wrote a 91,902-byte schema; `git diff --check` passed. |
 | [#39](https://github.com/lumensparkxy/agentic_test_case_generator/issues/39) | Extract frontend workflow state from `App.jsx` into clear domain hooks while preserving parse, refine, context, generate, integration, automation, execution, and export behavior. | `frontend/src/App.jsx`, `frontend/src/hooks/useWorkflowNavigationState.js`, `frontend/src/hooks/useRequirementWorkflowState.js`, `frontend/src/hooks/useContextWorkflowState.js`, `frontend/src/hooks/useTestCaseWorkflowState.js`, `frontend/src/hooks/useIntegrationWorkflowState.js`, `frontend/src/hooks/useExecutionWorkflowState.js`, `frontend/src/hooks/useExportWorkflowState.js`, `frontend/src/hooks/useAppSessionState.js`, `docs/codebase/ARCHITECTURE.md`, `docs/codebase/STRUCTURE.md` | `frontend/e2e/export-approval-gate.spec.js`, `frontend/e2e/workflow.spec.js`, `frontend/e2e/jira-workflow.spec.js` | `npm run build` passed; focused E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e -- e2e/export-approval-gate.spec.js` passed 1 test; full frontend E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e` passed 4 tests against local Vite and FastAPI with synthetic fixtures, including 40 generated cases with complete quality checks; backend suite passed 191 tests; strict requirement evaluation passed at 100/85; strict generation evaluation passed at 100/90; `scripts/export_openapi.py` wrote a 91,902-byte schema. |
 | [#40](https://github.com/lumensparkxy/agentic_test_case_generator/issues/40) | Modularize frontend CSS by feature ownership while preserving the existing workflow UI. | `frontend/src/App.jsx`, `frontend/src/styles/index.css`, `frontend/src/styles/layout.css`, `frontend/src/styles/forms.css`, `frontend/src/styles/workflow.css`, `frontend/src/styles/requirements-source.css`, `frontend/src/styles/integrations.css`, `frontend/src/styles/requirement-report.css`, `frontend/src/styles/test-cases.css`, `frontend/src/styles/feedback-context.css`, `frontend/src/styles/case-cards.css`, `frontend/src/styles/responsive.css`, `frontend/src/styles/export.css`, `frontend/src/styles/integration-cards.css`, `frontend/src/styles/generation-results.css`, `frontend/src/styles/generation-responsive.css`, `frontend/e2e/frontend-css-smoke.spec.js`, `docs/codebase/ARCHITECTURE.md`, `docs/codebase/STRUCTURE.md`, `docs/codebase/CONVENTIONS.md`, `docs/codebase/CONCERNS.md` | `frontend/e2e/frontend-css-smoke.spec.js`, `frontend/e2e/export-approval-gate.spec.js`, `frontend/e2e/workflow.spec.js`, `frontend/e2e/jira-workflow.spec.js` | `npm run build` passed; CSS smoke E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e -- e2e/frontend-css-smoke.spec.js` passed 4 desktop/mobile tests covering auth, settings, integrations, requirements, context, template, generate, automation, and export screens; focused export E2E passed 1 test; full frontend E2E `E2E_BASE_URL=http://127.0.0.1:5173 npm run test:e2e` passed 8 tests against local Vite and FastAPI with synthetic fixtures, including 40 generated cases with complete quality checks; backend suite passed 191 tests; strict requirement evaluation passed at 100/85; strict generation evaluation passed at 100/90; `scripts/export_openapi.py` wrote a 91,902-byte schema. |
+| [#41](https://github.com/lumensparkxy/agentic_test_case_generator/issues/41) | Split `backend/app/agents/test_case_agent.py` into focused backend modules while preserving public generation/refinement behavior and deterministic fallback coverage. | `backend/app/agents/test_case_agent.py`, `backend/app/agents/test_case_coverage.py`, `backend/app/agents/test_case_review.py`, `backend/app/agents/test_case_fallback.py`, `backend/app/agents/test_case_hydration.py`, `docs/codebase/ARCHITECTURE.md`, `docs/codebase/STRUCTURE.md`, `docs/codebase/CONCERNS.md`, `docs/implementation-plan.md` | `backend/tests/test_test_case_hydration.py`, `backend/tests/test_grounded_context_metrics.py`, `backend/tests/test_analysis_coverage_metrics.py`, `backend/tests/test_generation_recovery.py`, `backend/tests/test_review_retry.py`, `backend/tests/test_actor_propagation.py`, `backend/tests/test_observability_logging.py` | Focused backend tests passed 30 tests; backend suite `python -m unittest discover -s backend/tests -p 'test_*.py'` passed 191 tests; strict generation evaluation passed at 100/90 with unchanged benchmark coverage; strict requirement evaluation passed at 100/85; `scripts/export_openapi.py` wrote a 91,902-byte schema; generation-output E2E was not rerun because this was a behavior-preserving module split with no prompt, scoring, API path, or execution-candidate changes. |
